Create variuos templates and linked to navbar in base.html
