+++ .gitignore
@@ -1,0 +1,10 @@
+core.Microsoft*
+core.mongo*
+core.python*
+env.py
+__pycache__/
+*.py[cod]
+node_modules/
+
+# temporary ignored
+forms.py
