+++ requirements.txt
@@ -1,0 +1,8 @@
+click==8.0.1
+Flask==2.0.1
+Flask-WTF==0.15.1
+itsdangerous==2.0.1
+Jinja2==3.0.1
+MarkupSafe==2.0.1
+Werkzeug==2.0.1
+WTForms==2.3.3
