+++ static/js/script.js
@@ -1,4 +1,4 @@
 // Initialize Materialize with JQuery
 $(document).ready(function(){
-    $('.sidenav').sidenav();
+    $('.sidenav').sidenav({edge: "right"});
   });
