+++ templates/base.html
@@ -22,26 +22,29 @@
 </head>
 <body>
     <header>
+        <!-- top navbar -->
         <nav>
             <div class="nav-wrapper">
-            <a href="#!" class="brand-logo">EventPride</a>
-            <a href="#" data-target="mobile-demo" class="sidenav-trigger"><i class="material-icons">menu</i></a>
+            <a href="{{ url_for('index') }}" class="brand-logo">EventPride</a>
+            <a href="#" data-target="mobile-demo" class="sidenav-trigger right"><i class="fas fa-bars"></i></a>
             <ul class="right hide-on-med-and-down">
                 <li><a href="{{ url_for('index') }}">Home</a></li>
-                <li><a href="">Events</a></li>
-                <li><a href="">Wikipride</a></li>
-                <li><a href="">Login</a></li>
-                <li><a href="">Sign Up</a></li>
+                <li><a href="{{ url_for('get_events') }}">Events</a></li>
+                <li><a href="{{ url_for('wikipride') }}">Wikipride</a></li>
+                <li><a href="{{ url_for('login') }}">Login</a></li>
+                <li><a href="{{ url_for('signup') }}">Sign Up</a></li>
             </ul>
             </div>
         </nav>
 
+        <!-- mobile sidenav -->
         <ul class="sidenav" id="mobile-demo">
+            <li><h4 class="center-align">EventPride</h4></li>
             <li><a href="{{ url_for('index') }}">Home</a></li>
-            <li><a href="">Events</a></li>
-            <li><a href="">Wikipride</a></li>
-            <li><a href="">Login</a></li>
-            <li><a href="">Sign Up</a></li>
+            <li><a href="{{ url_for('get_events') }}">Events</a></li>
+            <li><a href="{{ url_for('wikipride') }}">Wikipride</a></li>
+            <li><a href="{{ url_for('login') }}">Login</a></li>
+            <li><a href="{{ url_for('signup') }}">Sign Up</a></li>
         </ul>
     </header>
