+++ templates/events.html
@@ -1,0 +1,5 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h2>Events</h2>
+{% endblock %}
