+++ templates/login.html
@@ -1,0 +1,5 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h2>Login</h2>
+{% endblock %}
