+++ templates/signup.html
@@ -1,0 +1,5 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h2>Sign Up</h2>
+{% endblock %}
