+++ templates/wikipride.html
@@ -1,0 +1,5 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <h2>Wikipride</h2>
+{% endblock %}
